rename batch->batchProcessing, fix comment, update reduce usage
