+++ src/batchProcessing.ts
