+++ src/lib/batchProcessing.ts
